revert: restaurar el firmware de la fase 4
El problema de la telemetria tenia otra causa, asi que se deshace el
revert anterior y vuelve el codigo completo: comandos de control, OTA,
historial en RTC RAM, traza circular, watchdog de RX y el fix de los ISR
de encoder.

+++ lib/motor_ctl_kalman/src/motor_ctl_kalman.cpp
@@ -198,8 +198,8 @@ void MotorController::update() {
 
   tickSampleTimePrev = tickTime;
 
-  long int encNow = getEncoderValue();
-  long int encDelta = encNow - encPrev;
+  int32_t encNow = getEncoderValue();
+  int32_t encDelta = encNow - encPrev;
   encPrev = encNow;
   float ticksPerMicroSec = ((float) encDelta) / ((float) tickTimeDelta);
   measuredRPM = ticksPerMicroSec * ticksPerMicroSecToRPM;
